+++ tests/func/mod.rs
@@ -1,2 +1,4 @@
 mod encode;
 mod decode;
+mod rserialize;
+mod serde;